refactor(domain): import variants from owner modules
