+++ src/flameox/application/capture.py
@@ -65,7 +65,6 @@ from flameox.domain import (
     CommandSpec,
     DomainError,
     ErrorCode,
-    ExecutionRunManifest,
     ExecutionStatus,
     ExternalExecutionContext,
     IdentityQuality,
@@ -81,7 +80,7 @@ from flameox.domain import (
     digest_model,
     new_id,
 )
-from flameox.domain.models import utc_now
+from flameox.domain.models import ExecutionRunManifest, utc_now
 from flameox.evidence import GenerationPublisher
 from flameox.execution import (
     ExecutionOutcome,
